+++ CHANGELOG.md
@@ -12,6 +12,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `emailTemplate` to `"PDF"` submission event
 - `FormsAppEnvironments` class
 
+### Breaking Changes
+
+- Updated `Forms.search()` to `Forms.searchForms()`.
+
 ## [0.16.0] - 2021-11-01
 
 ### Added